HashMapSolid
