+++ src/main/java/hashMap/MySortableConcurrentHashMapImpl.java
@@ -164,21 +164,6 @@ public class MySortableConcurrentHashMapImpl<K extends Comparable<K>, V> impleme
 
     @Override
     public Iterator<K> iterator() {
-        return new Iterator<K>() {
-
-            final List<K> keyList = getKeyList();
-            int i = 0;
-
-            @Override
-            public boolean hasNext() {
-                if (i < keyList.size()) return keyList.get(i) != null;
-                return false;
-            }
-
-            @Override
-            public K next() {
-                return keyList.get(i++);
-            }
-        };
+        return new MyIterator<K>(getKeyList());
     }
 }
